chore: move exception class to correct package
